+++ drivers/src/utils/mod.rs
@@ -1,7 +1,5 @@
 //! Event handler and device tree.
 
-#![allow(unused_imports)]
-
 mod event_listener;
 mod id_allocator;
 mod irq_manager;