+++ kernel-hal/src/bare/arch/riscv/interrupt.rs
@@ -1,5 +1,4 @@
 //! Interrupts management.
-
 use crate::{config::MAX_CORE_NUM, HalError, HalResult};
 use alloc::vec::Vec;
 use riscv::{asm, register::sstatus};
@@ -19,7 +18,10 @@ hal_fn_impl! {
 
         fn handle_irq(cause: usize) {
             trace!("Handle irq cause: {}", cause);
-            crate::drivers::all_irq().first_unwrap().handle_irq(cause)
+            let irq = crate::drivers::all_irq()
+                .find(alloc::format!("riscv-intc-cpu{}", crate::cpu::cpu_id()).as_str())
+                .expect("IRQ device 'riscv-intc' not initialized!");
+            irq.handle_irq(cause)
         }
 
         fn intr_on() {