+++ kernel-hal/src/bare/arch/x86_64/trap.rs
@@ -1,24 +1,20 @@
-#![allow(dead_code)]
-#![allow(clippy::identity_op)]
-
+use crate::context::TrapReason;
 use trapframe::TrapFrame;
 
-use crate::context::TrapReason;
-
 pub(super) const X86_INT_LOCAL_APIC_BASE: usize = 0xf0;
-pub(super) const X86_INT_APIC_SPURIOUS: usize = X86_INT_LOCAL_APIC_BASE + 0x0;
+pub(super) const _X86_INT_APIC_SPURIOUS: usize = X86_INT_LOCAL_APIC_BASE;
 pub(super) const X86_INT_APIC_TIMER: usize = X86_INT_LOCAL_APIC_BASE + 0x1;
-pub(super) const X86_INT_APIC_ERROR: usize = X86_INT_LOCAL_APIC_BASE + 0x2;
+pub(super) const _X86_INT_APIC_ERROR: usize = X86_INT_LOCAL_APIC_BASE + 0x2;
 
 // ISA IRQ numbers
-pub(super) const X86_ISA_IRQ_PIT: usize = 0;
-pub(super) const X86_ISA_IRQ_KEYBOARD: usize = 1;
-pub(super) const X86_ISA_IRQ_PIC2: usize = 2;
+pub(super) const _X86_ISA_IRQ_PIT: usize = 0;
+pub(super) const _X86_ISA_IRQ_KEYBOARD: usize = 1;
+pub(super) const _X86_ISA_IRQ_PIC2: usize = 2;
 pub(super) const X86_ISA_IRQ_COM2: usize = 3;
 pub(super) const X86_ISA_IRQ_COM1: usize = 4;
-pub(super) const X86_ISA_IRQ_CMOSRTC: usize = 8;
-pub(super) const X86_ISA_IRQ_MOUSE: usize = 12;
-pub(super) const X86_ISA_IRQ_IDE: usize = 14;
+pub(super) const _X86_ISA_IRQ_CMOSRTC: usize = 8;
+pub(super) const _X86_ISA_IRQ_MOUSE: usize = 12;
+pub(super) const _X86_ISA_IRQ_IDE: usize = 14;
 
 fn breakpoint() {
     panic!("\nEXCEPTION: Breakpoint");