+++ zCore/src/main.rs
@@ -1,9 +1,8 @@
-#![no_main]
 #![cfg_attr(not(feature = "libos"), no_std)]
-#![feature(lang_items)]
-#![feature(core_intrinsics)]
-#![feature(naked_functions, asm_sym, asm_const)]
 #![deny(warnings)]
+#![no_main]
+#![feature(naked_functions, asm_sym, asm_const)]
+#![feature(default_alloc_error_handler)]
 
 use core::sync::atomic::{AtomicBool, Ordering};